Rename loader module to logger module
